backend-flask/requirements.txt
aws-xray-sdk
flask
flask-cors
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp-proto-http 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-requests
watchtower
psycopg[binary]
psycopg[pool]